update workflow to run smoketest
