+++ .github/workflows/ci.yml
@@ -394,57 +394,3 @@ jobs:
           else
             echo "Integration tests passed"
           fi
-  smoke-test-released-exemplar:
-    name: Smoke Test Against Released Exemplar
-    if: github.event_name != 'pull_request' || github.event.action == 'enqueued' || contains( github.event.pull_request.labels.*.name, 'run-all')
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout the repository with full history
-        uses: actions/checkout@v4
-        with:
-          ref: ${{ github.head_ref }}
-          fetch-depth: 0
-          fetch-tags: true
-      - name: Get latest release tag
-        run: |
-          latest_release=$(git tag --sort=-v:refname | grep -v v1 | head -n 1)
-          echo "Latest release: $latest_release"
-          echo "LATEST_RELEASE=$latest_release" >> "$GITHUB_ENV"
-      - name: Check out exemplar from previous release
-        uses: actions/checkout@v4
-        with:
-          clean: false
-          repository: TBD54566975/ftl.git
-          path: tmpsmoketest
-          ref: ${{ env.LATEST_RELEASE }}
-          fetch-depth: 1
-          fetch-tags: true
-          sparse-checkout: smoketest
-          sparse-checkout-cone-mode: false
-      - name: Init Hermit
-        uses: cashapp/activate-hermit@v1
-        with:
-          cache: true
-      - name: Build Cache
-        uses: ./.github/actions/build-cache
-      - name: Docker Compose
-        run: docker compose up -d --wait
-      - name: Create DB
-        run: just init-db
-      - name: Download Go Modules
-        run: go mod download
-      - name: Smoke Test Upgrade Path
-        # There's a chicken-and-egg here since the smoketest tag isn't in the last release,
-        # update -tags integration to -tags smoketest when released.
-        run: |
-            set -euo pipefail
-            echo "Replacing current exemplar with latest released exemplar"
-            rm -r ./smoketest && mv ./tmpsmoketest/smoketest ./smoketest && rm -r ./tmpsmoketest
-            echo "Running smoke test"
-            go test -v -tags integration -run '^Test' ./smoketest
-      - name: Archive Report
-        uses: actions/upload-artifact@v4
-        if: always()
-        with:
-          name: kube-report-smoke-test-upgrade
-          path: /tmp/ftl-kube-report/
