+++ .github/workflows/smoketest.yml
@@ -59,24 +59,20 @@ jobs:
           cache: true
       - name: Build Cache
         uses: ./.github/actions/build-cache
-      - name: Docker Compose
-        run: docker compose up -d --wait
-      - name: Create DB
-        run: just init-db
       - name: Download Go Modules
         run: go mod download
       - name: Set up a kube cluster with the tagged release dockerhub image
         run: |
             set -euo pipefail
             echo "Deploying the tagged release to the cluster"
-            cd deployment && just deploy-version ${{ env.LATEST_VERSION }} && just wait-for-kube && cd ..
+            cd deployment && just deploy-version ${{ env.LATEST_VERSION }} && cd ..
       # We skip this since it requires an integration test change to skip the full kube deploy.
       # Re-enable this step when the tagged release has the integration test change.
       # - name: Smoke test the tagged release images
       #   run: |
       #       set -euo pipefail
       #       echo "Running smoke test on the tagged release images"
-      #       USE_DB_CONFIG=true SKIP_KUBE_FULL_DEPLOY=true go test -v -tags integration -run '^Test' ./smoketest
+      #       USE_DB_CONFIG=true SKIP_KUBE_FULL_DEPLOY=true go test -v -tags smoketest -run '^Test' ./smoketest
       - name: Check out HEAD of the current branch
         uses: actions/checkout@v4
         with:
@@ -86,7 +82,7 @@ jobs:
         run: |
             set -euo pipefail
             echo "Running smoke test on the HEAD images"
-            USE_DB_CONFIG=true go test -v -timeout 15m -tags integration -run '^Test' ./smoketest
+            USE_DB_CONFIG=true go test -v -timeout 15m -tags smoketest -run '^Test' ./smoketest
       - name: Archive Report
         uses: actions/upload-artifact@v4
         if: always()
